Entities "Cliente" and "Mascotas" created on H2. Relation between them ready.

+++ src/main/java/com/example/demo/servicio/ClienteServiceInterface.java
@@ -8,11 +8,11 @@ public interface ClienteServiceInterface {
 
     public boolean verifyCredentials(User user);
 
-    public Cliente SearchById(int id );
+    public Cliente SearchById(Long id );
 
     public Collection<Cliente> SearchAll();
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 
     public void update(Cliente cliente);
